pyproject.toml
[tool.poetry]
name = "hakaton"
version = "0.1.0"
description = ""
authors = ["alexeymyakinin <alexeymyakinin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.2"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
Markdown = "^3.4.3"
django-filter = "^23.2"
drf-yasg = "^1.21.6"
psycopg2-binary = "^2.9.6"
mimesis = "^10.1.0"
django-cors-headers = "^4.1.0"
aiogram = "2.25.1"
aiogram-dialog = "1.9.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
